% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_bp.R
\name{est_bp}
\alias{est_bp}
\title{Statistics of a BoxPlot}
\usage{
est_bp(boxp, date_factor, name_file)
}
\arguments{
\item{boxp}{Graph created with "grafica_BoxPlot"}

\item{date_factor}{Dates as a factor}

\item{name_file}{.xlsx file name}
}
\value{
Returns the quantiles 25, 50-25, 75-50 and the average of the data
}
\description{
The function "est_bp" Calculate the statistics of a boxplot
}
\references{
Ggplot2: Elegant Graphics for Data Analysis
}
\author{
Marcos Daniel Ramirez Santiago, \email{danielrs98@outlook.com}
}
