% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_BoxPlot.R
\name{grafica_BoxPlot}
\alias{grafica_BoxPlot}
\title{BoxPlot}
\usage{
grafica_BoxPlot(
  df,
  x_axis,
  y_axis,
  group_c,
  color_bp = colores[2],
  color_outlier = "#FDFEFE",
  legend_position = "right",
  legend_title = "",
  x_t = "",
  y_t = "",
  y_position = "right",
  formato_fecha = "\%m-\%Y",
  titulo = "",
  subtitulo = "",
  break_x = "8 week",
  t_hjust = 0,
  s_hjust = 0,
  f_wrap = F,
  f_wrap_1,
  scales_fw = T,
  bigotes = F
)
}
\arguments{
\item{df}{Default dataset to use for plot. If not already a data.frame, will be converted to one by fortify(). If not specified, must be supplied in each layer added to the plot.}

\item{x_axis}{data of variable x}

\item{y_axis}{variable to be plotted}

\item{group_c}{color fill}

\item{color_bp}{color of the BoxPlot from colores Banxico}

\item{color_outlier}{outlier}

\item{legend_position}{Position adjustment of the legend}

\item{legend_title}{Legend title to graph, as character}

\item{x_t}{x axis title to graph, as character}

\item{y_t}{y axis title to graph, as character}

\item{y_position}{Position of y axis, ie. ("right","left)}

\item{formato_fecha}{date format}

\item{titulo}{Title to graph, as character}

\item{subtitulo}{Subtitle to graph, as character}

\item{break_x}{x axis partitions}

\item{t_hjust}{Title position, values between [0,1]}

\item{s_hjust}{Subtitle position, values between [0,1]}

\item{f_wrap}{if f_wrap=T, then facet_wrap() function will be activate}

\item{f_wrap_1}{variables to graph}

\item{scales_fw}{scales of the variables graphed}

\item{bigotes}{T}
}
\value{
BoxPlot with Banxico format
}
\description{
The function "grafica_BoxPlot" create a BoxPlot for multiple series in a dataset
}
\references{
Ggplot2: Elegant Graphics for Data Analysis
}
\author{
Marcos Daniel Ramirez Santiago, \email{danielrs98@outlook.com}
}
