% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_pie.R
\name{grafica_pie}
\alias{grafica_pie}
\title{Pie Plot}
\usage{
grafica_pie(
  df,
  y_axis,
  fill_c,
  show_legend = T,
  etiquetas_c,
  i,
  legend_title = "",
  titulo,
  subtitulo
)
}
\arguments{
\item{df}{Default dataset to use for plot. If not already a data.frame, will be converted to one by fortify(). If not specified, must be supplied in each layer added to the plot.}

\item{y_axis}{variable to be plotted}

\item{fill_c}{color fill}

\item{show_legend}{shows the plotted data}

\item{etiquetas_c}{color of the data}

\item{i}{number of partitions of the pie}

\item{legend_title}{Legend title to graph, as character}

\item{titulo}{Title to graph, as character}

\item{subtitulo}{Subtitle to graph, as character}
}
\value{
BoxPlot with Banxico format
}
\description{
The function "grafica_pie" create a Pie Plot
}
\references{
Ggplot2: Elegant Graphics for Data Analysis
}
\author{
Marcos Daniel Ramirez Santiago, \email{danielrs98@outlook.com}
}
