% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica_ribbon.R
\name{grafica_ribbon}
\alias{grafica_ribbon}
\title{Line and Ribbon Plot}
\usage{
grafica_ribbon(
  df,
  x_axis,
  y_axis,
  fill_C,
  is_date = F,
  is_num = F,
  is_word = F,
  size_line = 1.5,
  size_point = 2,
  legend_position = "right",
  legend_title = "",
  x_t = "",
  y_t = "",
  y_position = "right",
  formato_fecha = "\%m-\%Y",
  titulo = "",
  subtitulo = "",
  break_x = "8 week",
  t_hjust = 0,
  s_hjust = 0,
  point = F,
  line_v = F,
  x_int,
  f_wrap = F,
  f_wrap_1,
  scales_fw = T,
  add_ribbon = F,
  y_axis_r,
  ribbon_col = "grey70",
  ribbon_alpha = 1,
  rint,
  size_t = 9,
  size_s = 9,
  size_x_t = 7,
  size_y_t = 7,
  size_l = 7,
  size_lt = 7,
  size_x_axis = 9,
  size_y_axis = 9,
  font_t = "Calibri",
  font_s = "Calibri",
  font_l = "Calibri",
  font_x = "Calibri",
  font_y = "Calibri",
  font_lt = "Calibri",
  col_t = colores[12],
  col_s = colores[12],
  angle_x = 0,
  angle_y = 0,
  x_axis_tf = "plain",
  y_axis_tf = "plain",
  x_axis_f = "plain",
  y_axis_f = "plain",
  line_alpha = 1,
  vline_alpha = 1,
  add_text = F,
  n_decimal = 2,
  size_text = 3.5,
  font_text = "plain",
  textp_y = 0.04,
  text_hjust = -0.25,
  escala = F,
  ay,
  by
)
}
\arguments{
\item{df}{Default dataset to use for plot. If not already a data.frame, will be converted to one by fortify(). If not specified, must be supplied in each layer added to the plot.}

\item{x_axis}{data of variable x}

\item{y_axis}{variable to be plotted}

\item{fill_C}{color fill}

\item{is_date}{if x-axis format are dates}

\item{is_num}{if x-axis format is numeric}

\item{is_word}{if x-axis format are words}

\item{size_line}{line size, numeric value}

\item{size_point}{point size, numeric value}

\item{legend_position}{Position adjustment of the legend}

\item{legend_title}{Legend title to graph, as character}

\item{x_t}{x axis title to graph, as character}

\item{y_t}{y axis title to graph, as character}

\item{y_position}{Position of y axis, ie. ("right","left)}

\item{formato_fecha}{date format}

\item{titulo}{Title to graph, as character}

\item{subtitulo}{Subtitle to graph, as character}

\item{break_x}{x axis partitions}

\item{t_hjust}{Title position, values between [0,1]}

\item{s_hjust}{Subtitle position, values between [0,1]}

\item{point}{If point = T, then add a point to the line graph}

\item{line_v}{if line_v=T, then add a vertical line}

\item{x_int}{Date where the vertical line will be placed}

\item{f_wrap}{if f_wrap=T, then facet_wrap() function will be activate}

\item{f_wrap_1}{variables to graph}

\item{scales_fw}{scales of the variables graphed}

\item{add_ribbon}{If add_ribbon = T, then ribbon format will be added to a time series}

\item{y_axis_r}{variable to be plotted for ribbon interval}

\item{ribbon_col}{color of the Ribbon area}

\item{ribbon_alpha}{ribbon opacity}

\item{rint}{Value that is subtracted from y_axis_r to obtain the interval defined by ymin and ymax}

\item{size_t}{title font size}

\item{size_s}{subtitle font size}

\item{size_x_t}{x-axis title font size}

\item{size_y_t}{y-axis title font size}

\item{size_l}{legend title font size}

\item{size_lt}{legend text font size}

\item{size_x_axis}{x-axis text font size}

\item{size_y_axis}{x-axis text font size}

\item{font_t}{title font style}

\item{font_s}{subtitle font style}

\item{font_l}{legend font style}

\item{font_x}{x-axis font style}

\item{font_y}{y-axis font style}

\item{font_lt}{legend text font style}

\item{col_t}{title color}

\item{col_s}{subtitle color}

\item{angle_x}{angle of x-axis text}

\item{angle_y}{angle of y-axis text}

\item{x_axis_tf}{title bold,plain, etc}

\item{y_axis_tf}{subtitle bold,plain, etc}

\item{x_axis_f}{bold,plain, etc}

\item{y_axis_f}{bold,plain, etc}

\item{line_alpha}{line plot opacity}

\item{vline_alpha}{vline opacity}

\item{add_text}{add text labels}

\item{n_decimal}{number of decimals to graph}

\item{size_text}{labels text font size}

\item{font_text}{labels font size}

\item{textp_y}{adjust text alignment}

\item{text_hjust}{adjust text alignment, values in [0,1]}

\item{escala}{adjust y-axis scale}

\item{ay}{y-axis scale lower limit}

\item{by}{y-axis scale upper limit}
}
\value{
Ribbon plot with Banxico format
}
\description{
The function "grafica_ribbon" For each x value, displays a "y" interval defined by ymin and ymax
}
\references{
Ggplot2: Elegant Graphics for Data Analysis
}
\author{
Marcos Daniel Ramirez Santiago, \email{danielrs98@outlook.com}
}
